requirements.txt
psycopg2-binary==2.8.6
python-ldap==3.3.1
kubernetes==17.17.0
simple-salesforce==1.11.1
pandas==1.1.5
retrying==1.3.3
